+++ R/product-rates.R
@@ -17,17 +17,9 @@
 #' from the 'uses' table in a [srppp_dm] object, as well as the columns
 #' 'percent' and 'g_per_L' from the 'ingredients' table in a [srppp_dm] object and
 #' 'culture_de' from the 'cultures' table in [srppp_dm] object.
-#' @param aggregation How to represent a range if present, e.g. "max" (default)
-#' or "mean".
-#' @param dosage_units If no units are given, or units are "%", then the applied
-#' amount in l/ha or kg/ha is calculated using a reference application volume and
-#' the dosage. As the dosage units are not explicitly given, we can specify our
-#' assumptions about these using this argument (currently not implemented, i.e.
-#' specifying the argument has no effect).
-#' @param fix_l_per_ha During the review of the 2023 indicator
-#' project calculations, a number of cases were identified where the unit
-#' l/ha specifies a water volume, and not a product volume. If TRUE (default),
-#' these cases are corrected, if FALSE, these cases are discarded.
+#' @param skip_l_per_ha_without_g_per_L Passed on to `application_rate_g_per_ha`,
+#' but here, the default is FALSE, there it is TRUE.
+#' @inheritParams application_rate_g_per_ha
 #' @return A tibble containing one additional column 'rate_g_per_ha'
 #' @export
 #' @examples
@@ -56,7 +48,7 @@
 #' }
 product_rates <- function(product_uses,
   aggregation = c("max", "mean", "min"),
-  dosage_units = c("percent_ww", "percent_vv", "state_of_matter"),
+  skip_l_per_ha_without_g_per_L = FALSE,
   fix_l_per_ha = TRUE)
 {
   aggregation = match.arg(aggregation)
@@ -75,19 +67,22 @@ product_rates <- function(product_uses,
         min_dosage, max_dosage)
       ) |>
     mutate(
-      rate = case_when(
-        aggregation == "mean" ~ (rate_min + rate_max)/2,
-        aggregation == "max" ~ rate_max,
-        aggregation == "min" ~ rate_min
-      ),
-      dosage = case_when(
-        aggregation == "mean" ~ (dosage_min + dosage_max)/2,
-        aggregation == "max" ~ dosage_max,
-        aggregation == "min" ~ dosage_min
-      )
+      rate = if (aggregation == "mean") {
+        (rate_min + rate_max)/2
+      } else if (aggregation == "max") {
+        rate_max
+      } else {
+        rate_min
+      },
+      dosage = if (aggregation == "mean") {
+        (dosage_min + dosage_max)/2
+      } else if (aggregation == "max") {
+        dosage_max
+      } else {
+        dosage_min
+      }
     )
 
-  # if (is.na(rates_dosages$culture_de))
   product_rates <- rates_dosages |>
     mutate(ref_volume = case_when(
       application_area_de %in% c("Feldbau", "Gem\u00FCsebau", "Beerenbau", "Zierpflanzen")
@@ -98,20 +93,20 @@ product_rates <- function(product_uses,
     left_join(l_per_ha_is_water_volume, by = c("pNbr", "use_nr")) |>
     mutate(prod_rate = case_when(
       units_de == "l/ha" ~ # l/ha can refer to product or water volume
-        if_else(is.na(source), # if no external information, assume l/ha is product
-                if_else(is.na(g_per_L) & !is.na(percent), # if g_per_L is not defined and percent is given
-                        if_else(!is.na(dosage) & rate > 100,
-                                # If we have a dosage and the rate is above 100 l/ha, the rate in l/ha is assumed to be the application solution
-                                rate * dosage / 100, # Correct for Rhodofix 2009 (Grünbuch) and 2012 (XML)
-                                # If we have no dosage, treat l/ha as kg/ha product rate
-                                rate # Correct for Metro 2017
-
-                                ),
-                        rate # l/ha is product
-                ),
-                if (fix_l_per_ha) {
-                  rate * dosage / 100 # external info: l/ha for solution, weighted by dosage
-                } else NA),
+        if_else(is.na(source) | rate < 100, # if no external information, or the rate is < 100 l/ha, start assuming that l/ha is product
+          if_else(is.na(g_per_L), # if g_per_L is not defined
+            if (skip_l_per_ha_without_g_per_L) NA # set to NA, as in the 2023 indicator
+            else {
+              if_else(is.na(dosage),
+                # If we have no dosage, treat l/ha as kg/ha product rate
+                rate, # l/ha is assumed to be product, correct for Metro 2017
+                rate * dosage/100) # else assume that the rate refers to the application solution, and use dosage information (e.g. the corrected dosage for Rhodofix, or the amended dosage for Dirigol - N)
+            },
+            rate
+          ),
+          if (fix_l_per_ha) {
+            rate * dosage / 100 # external info: l/ha is solution, use dosage
+          } else NA),
       units_de == "kg/ha" ~ rate,                     # product already in kg/ha
       units_de == "g/ha" ~ rate / 1000,               # convert g/ha → kg/ha
       units_de == "ml/m\u00B2" ~ rate / 1000 * 10000, # ml/m² → L/ha
@@ -130,9 +125,15 @@ product_rates <- function(product_uses,
   return(ret)
 }
 
-
 #' Use definitions where the rate in l/ha refers to the volume of the spraying solution
 #'
+#' This information is used in the functions [product_rates] and [application_rate_g_per_ha]
+#' in cases where a rate in l/ha exceeds 100 l/ha. It only affects older XML files,
+#' in current versions of the XML files, rate specifications always refer to
+#' the product.
+#'
+#' Currently, the affected products are Dormex (P-Nr. 5151) and Karate with
+#' Zeon technology (P-Nr. 3756).
 #' @docType data
 #' @export
 #' @seealso [product_rates]
@@ -140,7 +141,9 @@ product_rates <- function(product_uses,
 #' library(srppp)
 #' l_per_ha_is_water_volume
 l_per_ha_is_water_volume <- tibble::tribble(
-  ~ pNbr, ~ use_nr, ~ source, ~ url,
+  ~ pNbr, ~ use_nr, ~ source, ~ url, ~ file,
   5151L, 1L, "EFSA conclusion on cyanamide 2010, p. 17",
-  "https://doi.org/10.2903/j.efsa.2010.1873"
+  "https://doi.org/10.2903/j.efsa.2010.1873", NA,
+  3756L, 14L, "Verzeichnis 2009 Pflanzenschutzmittel",
+  NA, "Gr\u00FCnbuch 20090518.pdf"
 )